# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for the BONUSPLAY demo rewards platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v2/rewards:
    get:
      operationId: listV2Rewards
      responses:
        "200":
          description: Approved rewards and closed activation gate
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2RewardPage" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/rewards/{rewardId}:
    get:
      operationId: getV2Reward
      parameters:
        - { name: rewardId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Reward detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2RewardDetail" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/rewards/{rewardId}/redeem:
    post:
      operationId: redeemV2Reward
      parameters:
        - { name: rewardId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2RedemptionInput" }
      responses:
        "200":
          description: Reserved order or current replay
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2RewardOrder" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/orders:
    get:
      operationId: listV2Orders
      responses:
        "200":
          description: Own orders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2OrderPage" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/reward-suppliers:
    get:
      operationId: listV2RewardSuppliers
      responses:
        "200":
          description: Operator configured approved suppliers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/V2RewardSupplier" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/rewards:
    get:
      operationId: listV2AdminRewards
      responses:
        "200":
          description: All reward definitions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2RewardPage" }
        default: { $ref: "#/components/responses/V2PointsError" }
    post:
      operationId: createV2Reward
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2RewardInput" }
      responses:
        "201":
          description: Immutable draft
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2Reward" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/rewards/{rewardId}/review:
    post:
      operationId: reviewV2Reward
      parameters:
        - { name: rewardId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2RewardReviewInput" }
      responses:
        "200":
          description: Reviewed reward
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2Reward" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/orders:
    get:
      operationId: listV2AdminOrders
      responses:
        "200":
          description: Delivery queue and history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2OrderPage" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/orders/{orderId}/action:
    post:
      operationId: actionV2Order
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2OrderActionInput" }
      responses:
        "200":
          description: Delivery action or current replay
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2RewardOrder" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/rewards/reconciliation:
    get:
      operationId: reconcileV2Rewards
      responses:
        "200":
          description: Read only stock and ledger reconciliation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2RewardReconciliation" }
        default: { $ref: "#/components/responses/V2PointsError" }
...
  /v2/wallet:
    get:
      operationId: getV2Wallet
      tags: [v2]
      summary: Read own real ledger balance, never demo balances
      responses:
        "200":
          description: Points wallet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2Wallet" }
        default:
          $ref: "#/components/responses/V2PointsError"
  /v2/transactions:
    get:
      operationId: listV2Transactions
      tags: [v2]
      parameters:
        - $ref: "#/components/parameters/PointsCursor"
        - $ref: "#/components/parameters/PointsLimit"
      responses:
        "200":
          description: Newest first keyset-paginated history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2TransactionPage" }
        default:
          $ref: "#/components/responses/V2PointsError"
  /v2/admin/accounts/{accountId}/wallet:
    get:
      operationId: getV2AdminWallet
      tags: [v2]
      parameters:
        - $ref: "#/components/parameters/PointsAccountId"
      responses:
        "200":
          description: Administrator account wallet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2Wallet" }
        default:
          $ref: "#/components/responses/V2PointsError"
  /v2/admin/accounts/{accountId}/transactions:
    get:
      operationId: listV2AdminTransactions
      tags: [v2]
      parameters:
        - $ref: "#/components/parameters/PointsAccountId"
        - $ref: "#/components/parameters/PointsCursor"
        - $ref: "#/components/parameters/PointsLimit"
      responses:
        "200":
          description: Administrator account history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2TransactionPage" }
        default:
          $ref: "#/components/responses/V2PointsError"
  /v2/admin/accounts/{accountId}/adjustments:
    post:
      operationId: adjustV2Points
      tags: [v2]
      summary: Audited administrator adjustment with mandatory reason
      parameters:
        - $ref: "#/components/parameters/PointsAccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2PointsAdjustment" }
      responses:
        "202":
          description: High-risk request recorded; executes only after the cooldown and an explicit MFA-verified confirmation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2AdminRequest" }
        default:
          $ref: "#/components/responses/V2PointsError"
  /v2/admin/transactions/{transactionId}/decision:
    post:
      operationId: decideV2Points
      tags: [v2]
      summary: Approve or reject a pending transaction, append-only
      parameters:
        - $ref: "#/components/parameters/PointsTransactionId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2PointsDecision" }
      responses:
        "202":
          description: High-risk request recorded; executes only after the cooldown and an explicit MFA-verified confirmation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2AdminRequest" }
        default:
          $ref: "#/components/responses/V2PointsError"
  /v2/admin/transactions/{transactionId}/compensation:
    post:
      operationId: compensateV2Points
      tags: [v2]
      summary: Full refund of a debit or full reversal of a credit/debit
      parameters:
        - $ref: "#/components/parameters/PointsTransactionId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2PointsCompensation" }
      responses:
        "202":
          description: High-risk request recorded; executes only after the cooldown and an explicit MFA-verified confirmation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2AdminRequest" }
        default:
          $ref: "#/components/responses/V2PointsError"
  /v2/me:
    get:
      operationId: getV2Account
      tags: [v2]
      summary: Get the authenticated V2 profile, or null before enrollment
      responses:
        "200":
          description: Account state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2AccountState" }
        "401":
          description: Login required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      operationId: updateV2Account
      tags: [v2]
      summary: Update the current account profile, never roles or balances
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2ProfileInput" }
      responses:
        "200":
          description: Updated account state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2AccountState" }
  /v2/enroll:
    post:
      operationId: enrollV2Account
      tags: [v2]
      summary: Enroll a verified identity with versioned legal acceptance
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2EnrollmentInput" }
      responses:
        "201":
          description: Account enrolled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2AccountState" }
  /v2/admin/access:
    get:
      operationId: checkV2AdminAccess
      tags: [v2]
      summary: Server-side administrator authorization probe
      responses:
        "200":
          description: Access permitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2Access" }
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /v2/partner/access:
    get:
      operationId: checkV2PartnerAccess
      tags: [v2]
      summary: Server-side partner authorization probe
      responses:
        "200":
          description: Access permitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2Access" }
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  responses:
    V2PointsError:
      description: 400 invalid input; 401 login required; 403 forbidden; 404 not found; 409 replay conflict, insufficient balance or invalid transition; 429 rate limit
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  parameters:
...
    PointsAccountId:
      name: accountId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
    PointsTransactionId:
      name: transactionId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PointsCursor:
      name: cursor
      in: query
      schema: { type: string, pattern: '^[1-9][0-9]{0,18}$' }
    PointsLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  schemas:
...
    V2AdminRequest:
      type: object
      required: [id, action, payload, reason, requestedBy, requestedByName, targetAccountId, requestedAt, notBefore, status, decidedBy, decidedAt, note, result, canConfirm, blockedReason]
      properties:
        id: { type: string }
        action: { type: string, enum: [POINTS_ADJUSTMENT, POINTS_DECISION, POINTS_COMPENSATION, OFFER_APPROVAL, REWARD_APPROVAL, ORDER_DISPATCH, ECONOMY_CONFIG, MARKETING_BUDGET] }
        payload: { type: object, additionalProperties: true }
        reason: { type: string }
        requestedBy: { type: string }
        requestedByName: { type: string }
        targetAccountId: { type: [string, "null"] }
        requestedAt: { type: string }
        notBefore: { type: string }
        status: { type: string, enum: [pending, executed, rejected, cancelled] }
        decidedBy: { type: [string, "null"] }
        decidedAt: { type: [string, "null"] }
        note: { type: [string, "null"] }
        result: { type: [object, "null"], additionalProperties: true }
        canConfirm: { type: boolean }
        blockedReason: { type: [string, "null"] }
...
    V2RewardSupplier:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
    V2RewardInput:
      type: object
      required: [supplierId, title, description, terms, points, stock, supplierSku, approvalReference]
      properties:
        supplierId: { type: string, minLength: 3, maxLength: 40 }
        title: { type: string, minLength: 3, maxLength: 120 }
        description: { type: string, minLength: 10, maxLength: 2000 }
        terms: { type: string, minLength: 10, maxLength: 2000 }
        points: { type: integer, minimum: 1, maximum: 1000000 }
        stock: { type: integer, minimum: 0, maximum: 1000000 }
        supplierSku: { type: string, minLength: 1, maxLength: 128 }
        approvalReference: { type: string, minLength: 10, maxLength: 200 }
    V2Reward:
      type: object
      required: [id, supplierId, supplierName, title, description, terms, points, stock, status]
      properties:
        id: { type: string, format: uuid }
        supplierId: { type: string }
        supplierName: { type: string }
        title: { type: string }
        description: { type: string }
        terms: { type: string }
        points: { type: integer }
        stock: { type: integer }
        status: { type: string, enum: [draft, approved, disabled] }
    V2RewardPage:
      type: object
      required: [items, redeemEnabled]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/V2Reward" } }
        redeemEnabled: { type: boolean }
    V2RewardDetail:
      type: object
      required: [reward, redeemEnabled]
      properties:
        reward: { $ref: "#/components/schemas/V2Reward" }
        redeemEnabled: { type: boolean }
    V2RedemptionInput:
      type: object
      required: [idempotencyKey]
      properties:
        idempotencyKey: { type: string, format: uuid }
    V2RewardReviewInput:
      type: object
      required: [status, reason]
      properties:
        status: { type: string, enum: [approved, disabled] }
        reason: { type: string, minLength: 10, maxLength: 500 }
    V2RewardOrder:
      type: object
      required: [id, rewardId, title, points, status, transactionId, createdAt, updatedAt]
      properties:
        accountId: { type: string }
        supplierId: { type: string }
        supplierSku: { type: string }
        id: { type: string, format: uuid }
        rewardId: { type: string, format: uuid }
        title: { type: string }
        points: { type: integer }
        status: { type: string, enum: [reserved, delivering, uncertain, delivered, refunded] }
        transactionId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    V2OrderPage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/V2RewardOrder" } }
    V2OrderActionInput:
      type: object
      required: [action, reason, evidenceReference, idempotencyKey]
      properties:
        confirmedNotDelivered: { type: boolean }
        action: { type: string, enum: [dispatch, delivered, uncertain, refund] }
        reason: { type: string, minLength: 10, maxLength: 500 }
        evidenceReference: { type: string, minLength: 10, maxLength: 200 }
        idempotencyKey: { type: string, format: uuid }
    V2RewardReconciliation:
      type: object
      required: [ok, issues]
      properties:
        ok: { type: boolean }
        issues:
          type: array
          items:
            type: object
            required: [code, reference]
            properties:
              code: { type: string }
              reference: { type: string }
...
    V2PointsType:
      type: string
      enum: [EARN, REDEEM, REFERRAL, BONUS, ADJUSTMENT, REFUND, REVERSAL, EXPIRATION]
    V2PointsStatus:
      type: string
      enum: [pending, approved, rejected, reversed]
    V2Wallet:
      type: object
      required: [accountId, available, pending, reserved, balance, lifetimeEarned, lifetimeRedeemed]
      properties:
        accountId: { type: string }
        available: { type: integer, minimum: 0 }
        pending: { type: integer, minimum: 0 }
        reserved: { type: integer, minimum: 0 }
        balance: { type: integer, minimum: 0 }
        lifetimeEarned: { type: integer, minimum: 0 }
        lifetimeRedeemed: { type: integer, minimum: 0 }
    V2PointsTransaction:
      type: object
      required: [id, sequence, accountId, type, amount, status, source, reference, description, reason, relatedTransactionId, actorId, createdAt, updatedAt, events]
      properties:
        id: { type: string, format: uuid }
        sequence: { type: string }
        accountId: { type: string }
        type: { $ref: "#/components/schemas/V2PointsType" }
        amount: { type: integer }
        status: { $ref: "#/components/schemas/V2PointsStatus" }
        source: { type: string }
        reference: { type: string }
        description: { type: string }
        reason: { type: string }
        relatedTransactionId: { type: [string, "null"] }
        actorId: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
        events:
          type: array
          items: { $ref: "#/components/schemas/V2PointsEvent" }
    V2PointsEvent:
      type: object
      required: [status, delta, reservedDelta, reason, actorId, createdAt]
      properties:
        status: { $ref: "#/components/schemas/V2PointsStatus" }
        delta: { type: integer }
        reservedDelta: { type: integer }
        reason: { type: string }
        actorId: { type: string }
        createdAt: { type: string }
    V2TransactionPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/V2PointsTransaction" }
        nextCursor: { type: [string, "null"] }
    V2PointsAdjustment:
      type: object
      additionalProperties: false
      required: [amount, reason, idempotencyKey]
      properties:
        amount: { type: integer, minimum: -1000000, maximum: 1000000 }
        reason: { type: string, minLength: 10, maxLength: 500 }
        idempotencyKey: { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9_.:-]{16,128}$' }
    V2PointsDecision:
      type: object
      additionalProperties: false
      required: [status, reason, idempotencyKey]
      properties:
        status: { type: string, enum: [approved, rejected] }
        reason: { type: string, minLength: 10, maxLength: 500 }
        idempotencyKey: { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9_.:-]{16,128}$' }
    V2PointsCompensation:
      type: object
      additionalProperties: false
      required: [type, reason, idempotencyKey]
      properties:
        type: { type: string, enum: [REFUND, REVERSAL] }
        reason: { type: string, minLength: 10, maxLength: 500 }
        idempotencyKey: { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9_.:-]{16,128}$' }
...
    V2ProfileInput:
      type: object
      additionalProperties: false
      required: [firstName, lastName, country, language]
      properties:
        firstName: { type: string, minLength: 1, maxLength: 80 }
        lastName: { type: string, minLength: 1, maxLength: 80 }
        country: { type: string, enum: [NO] }
        language: { type: string, enum: [nb, en] }
    V2EnrollmentInput:
      type: object
      additionalProperties: false
      required: [firstName, lastName, country, language, termsAccepted, privacyAccepted, termsVersion, privacyVersion, ageConfirmed]
      properties:
        firstName: { type: string, minLength: 1, maxLength: 80 }
        lastName: { type: string, minLength: 1, maxLength: 80 }
        country: { type: string, enum: [NO] }
        language: { type: string, enum: [nb, en] }
        termsAccepted: { type: boolean, enum: [true] }
        privacyAccepted: { type: boolean, enum: [true] }
        ageConfirmed: { type: boolean, enum: [true] }
        termsVersion: { type: string, maxLength: 64 }
        privacyVersion: { type: string, maxLength: 64 }
    V2Account:
      type: object
      required: [id, email, firstName, lastName, country, language, role, status, referralCode, termsVersion, privacyVersion, termsAcceptedAt, privacyAcceptedAt, emailVerifiedAt, createdAt, lastLoginAt]
      properties:
        id: { type: string }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        country: { type: string }
        language: { type: string }
        role: { type: string, enum: [USER, PARTNER, ADMIN, SUPER_ADMIN] }
        status: { type: string, enum: [ACTIVE, SUSPENDED, DELETION_REQUESTED] }
        referralCode: { type: string }
        termsVersion: { type: string }
        privacyVersion: { type: string }
        termsAcceptedAt: { type: string }
        privacyAcceptedAt: { type: string }
        emailVerifiedAt: { type: string }
        createdAt: { type: string }
        lastLoginAt: { type: string }
    V2AccountState:
      type: object
      required: [account, termsVersion, privacyVersion, phase, commerceEnabled]
      properties:
        account:
          anyOf:
            - { $ref: "#/components/schemas/V2Account" }
            - { type: "null" }
        termsVersion: { type: string }
        privacyVersion: { type: string }
        phase: { type: string, enum: [FOUNDATION, POINTS] }
        commerceEnabled: { type: boolean, enum: [false] }
    V2Access:
      type: object
      required: [ok, role]
      properties:
        ok: { type: boolean }
        role: { type: string, enum: [USER, PARTNER, ADMIN, SUPER_ADMIN] }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...